feat: Break up package lists for manageability
This is a *huge* change.  It break package lists in to a file per
section to keep the size a little more under control.  I’ve filtered the
repository to apply the change to all the previous commits too.

I still expect some breakages from this, as it hasn’t really been tested
to thoroughly.  Let’s see what happens…